Ajout du ballon de basket pour faire fuir les fantomes !

+++ Ez_Maze_Git/core/src/com/mygdx/ezmaze/jeu/objects/Fantome.java
@@ -39,7 +39,7 @@ public class Fantome extends AbstractGameObject {
 	
 	}
 
-	public void init(int x, int y) {
+	public void init(float x, float y) {
 		dimension.set(0.5f,0.5f);
 		regFantome = Assets.instance.monstre.fantome;
